WIP: remove stream read json file and refactoring.
