+++ config/config.dev.yml
@@ -2,7 +2,7 @@ redis:
   hostname: '127.0.0.1'
   port: 6379
   password: ''
-  database: 0
+  database: 15
 
 webserver:
   port: ':9001'
