+++ main.go
@@ -3,8 +3,9 @@ package main
 import (
 	configLoader "app/config"
 	"app/providers"
-	"app/redis"
-	"app/webserver"
+	"app/server"
+	"app/storage/redis"
+	"fmt"
 	"log"
 )
 
@@ -12,11 +13,12 @@ func main() {
 
 	var config = configLoader.Load()
 
-	log.Println("redis hostname: ", config.Redis.Hostname)
-
 	log.Println("Start web server")
 	log.Println(config.WebServer.Port)
-	providers.SelectStrategyDownloadFile("dsa")
+
 	redis.Connect(config.Redis)
-	webserver.Start(config.WebServer)
+
+	test := providers.StrategyFactoryProvider("resources/data/data.json")
+	fmt.Println(test.Load())
+	server.Start(config.WebServer)
 }
