+++ models/parkingTaxi.go
@@ -1,5 +1,7 @@
 package models
 
+var AllowedIndex = []string{"GlobalID", "ID", "Mode"}
+
 type ParkingTaxi struct {
 	GlobalID              int    `json:"global_id"`
 	SystemObjectID        string `json:"system_object_id"`
