+++ storage/redis/storage.go
@@ -2,7 +2,6 @@ package redis
 
 import (
 	"app/config"
-	"fmt"
 	"github.com/go-redis/redis/v7"
 	"log"
 	"strconv"
@@ -12,21 +11,21 @@ var client *redis.Client
 
 func Connect(config config.RedisConfiguration) *redis.Client {
 
-	client = redis.NewClient(&redis.Options{
+	redisClient := redis.NewClient(&redis.Options{
 		Addr:     config.Hostname + ":" + strconv.Itoa(config.Port),
 		Password: config.Password,
 		DB:       config.Database,
 	})
 
-	pong, err := client.Ping().Result()
+	_, err := redisClient.Ping().Result()
 
 	if err != nil {
 		log.Fatal("i can not connected to Redis. Check config file.", err)
 	}
 
-	fmt.Println(pong)
+	client = redisClient
 
-	return client
+	return redisClient
 }
 
 func GetDB() *redis.Client {
