+++ storage/storage.go
@@ -1,0 +1,4 @@
+package storage
+
+type Storage interface {
+}
